// src/deltas/domain/enums.ts
/**
 * Auto generated. DO NOT edit manually.
 * Last updated on: Thu, 29 Jun 2023 08:50:21 GMT
 */

/**
 * @enum
 * @readonly
 * @description TokenWorlds Table Names
 */
export enum TokenWorldsTableName {
    Accounts = 'accounts',
    Members = 'members',
    Memberterms = 'memberterms',
    Stakeconfig = 'stakeconfig',
    Stakes = 'stakes',
    Staketime = 'staketime',
    Stat = 'stat',
    Unstakes = 'unstakes',
};